+++ resources/js/app.js
@@ -1,8 +1,10 @@
 import Alpine from 'alpinejs';
 import intersect from '@alpinejs/intersect';
+import registerLandingHome from './landing-home';
 import './cookie-banner';
 
 Alpine.plugin(intersect);
+registerLandingHome(Alpine);
 
 window.Alpine = Alpine;